feat(timetable): add getTimetable by userId

+++ heyyoung/src/main/java/com/alddeul/heyyoung/timetable/dto/UserTimeTableDto.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class UserTimeTableDto {
 
     public static record Response(
-            int userId,
+    		long userId,
             List<LectureDto> lectures
     ) {}
